+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+# [0.53.1] - 2024-05-10
+
+### Added
+- `analysis disable` and `analysis enable` commands added;
+- `analysis list` can now optionally show disabled analyses
+
 # [0.53.0] - 2024-05-01
 ### Changed
 - `seq purge` by default purges only inactive sequences for a given sample.